refactor: BeanFactory 적용 리팩터링
